backend/requirements.txt
fastapi
uvicorn
transformers
torch
Pillow
sqlite3
speechrecognition
